+++ shared/core/src/commonMain/kotlin/com/loc/hexis/core/tasks/PomodoroRepo.kt
@@ -16,6 +16,8 @@ interface PomodoroRepo {
 
     suspend fun getTodayStats(): PomodoroStats
 
+    fun getTodayStatsFlow(): Flow<PomodoroStats>
+
     fun getCompletedDates(): Flow<List<LocalDate>>
 
     suspend fun getEarliestSessionDate(): LocalDate?